Warning stomping and some code cleanup in GMGen:  Add Generics  Change for loops to Java 1.5 loops (mainly for generics)  Eliminate assignment of parameters  Use int for calculations, not Integer  Add some TODOs to trigger some reviews of unused or potentially updateable code     (MutableFilter.java, PcgCombatant.java, GMGenSystemView.java, ErrorListDialog.java, etc.)  Use Integer.toString() rather than int + ""  Use CollectionUtilities.joinStringRepresentations() where possible  Make inner classes static where possible  Eliminate unused variables  Eliminate useless null checks  Make constants static where possible  Change bitwise ORs | in IFs to logical ORs ||
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@1557 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/gmgen/plugin/Dice.java
@@ -42,7 +42,7 @@ public class Dice extends Die
 		this.num = num;
 		this.sides = sides;
 		this.aModifier = modifier;
-		rolls = new Integer[num];
+		rolls = new int[num];
 		roll();
 	}
 
@@ -67,8 +67,8 @@ public class Dice extends Die
 
 		for (i = 0; i < num; i++)
 		{
-			rolls[i] = Integer.valueOf(rand.nextInt(sides) + 1);
-			value = rolls[i].intValue() + value;
+			rolls[i] = rand.nextInt(sides) + 1;
+			value = rolls[i] + value;
 		}
 
 		total = value + aModifier;
